+++ trunk/vcf/src/vcf/ApplicationKit/Win32Dialog.cpp
@@ -288,7 +288,7 @@ void Win32Dialog::showMessage( const String& message, const String& caption )
 }
 
 UIToolkit::ModalReturnType Win32Dialog::showMessage( const String& message, const String& caption,
-												 const long& messageButtons, const Dialog::MessageStyle& messageStyle )
+												 const int32& messageButtons, const Dialog::MessageStyle& messageStyle )
 {
 	UIToolkit::ModalReturnType result = UIToolkit::mrNone;
