+++ trunk/vcf/src/vcf/ApplicationKit/Win32MenuItem.h
@@ -38,11 +38,11 @@ public:
 
 	virtual void addChild( MenuItem* child );
 
-	virtual void insertChild( const unsigned long& index, MenuItem* child );
+	virtual void insertChild( const uint32& index, MenuItem* child );
 
 	virtual void deleteChild( MenuItem* child );
 
-	virtual void deleteChild( const unsigned long& index );
+	virtual void deleteChild( const uint32& index );
 
 	virtual void clearChildren();
 
@@ -54,7 +54,7 @@ public:
 
 	virtual bool hasParent();
 
-	virtual MenuItem* getChildAt( const unsigned long& index );
+	virtual MenuItem* getChildAt( const uint32& index );
 
 	virtual bool isEnabled();
